optimized using float32 and compilation fixed. need to delete time printings

+++ MovingCameraForegroundEstimetor/KLTWrapper.py
@@ -54,10 +54,11 @@ class KLTWrapper:
         if self.count > 10:
             self.makeHomoGraphy(good1, good2)
         else:
-            self.H = np.identity(3)
+            self.H = np.identity(3, dtype=np.float32)
         self.InitFeatures(image)
         self.imgPrevGray = image
         return self.H
 
     def makeHomoGraphy(self, p1, p2):
         self.H, status = cv2.findHomography(p1, p2, cv2.RANSAC, 1.0)
+        self.H = self.H.astype(np.float32)
